+++ qp_klp/klp.py
@@ -29,12 +29,20 @@ class StatusUpdate():
         # internal function implements a callback function for Pipeline.run().
         # :param id: PBS/Torque/or some other informative and current job id.
         # :param status: status message
+
+        # Note that jid and self.job_id are not necessarily the same. jid is
+        # the job id of whatever child process is currently running, while
+        # self.job_id is the job id for the main, master job.
         self.qclient.update_job_step(self.job_id,
                                      self.msg + f" ({jid}: {status})")
 
     def update_current_message(self, step=None):
         # internal function that sets current_message to the new value before
         # updating the job step in the UI.
+
+        # Note that unlike update_job_status, this function is not meant to be
+        # used as a callback function. It is meant to be called directly by
+        # the plugin and whatever else is passed this object.
         if step:
             msg = "Step %d of %d: " % (step[0], step[1])
         else:
@@ -64,24 +72,35 @@ def sequence_processing_pipeline(qclient, job_id, parameters, out_dir):
     bool, list, str
         The results of the job
     """
+    status_line = StatusUpdate(qclient, job_id)
 
-    # if a file exists in the working (out) directory named, 'restart_me'
-    # then consider this a job that executed previously but failed.
+    # NB: RESTART FUNCTIONALITY:
+    # To restart a job that's failed, simply create a ProcessingJob() object
+    # in Qiita's interpreter using the Qiita Job ID aka the name of the
+    # working directory. _set_status("in construction") and then call
+    # submit(). New SPP jobs will record all the information needed to perform
+    # a restart should it fail.
     out_path = partial(join, out_dir)
     is_restart = True if exists(out_path('restart_me')) else False
 
-    status_line = StatusUpdate(qclient, job_id)
-
     if is_restart:
         status_line.update_current_message("Restarting pipeline")
-        # Assume ConvertJob directory exists and parse the job-script found
-        # there. If this is a restart, we won't be given the run-identifier,
-        # the lane number, and the sample-sheet as input parameters.
-        some_path = join(out_dir, 'ConvertJob', 'ConvertJob.sh')
-        result = ConvertJob.parse_job_script(some_path)
-        run_identifier = split(result['run_directory'])[-1]
-        uif_path = result['sample_sheet_path']
+        with open(out_path('restart_me'), 'r') as f:
+            lines = f.readlines()
+            lines = [x.strip() for x in lines]
+            lines = [x for x in lines if x != '']
+            if len(lines) != 2:
+                raise ValueError(f"{out_path('restart_me')} can only contain "
+                                 "the run-identifier on line 1 and the name "
+                                 "of the user input file on line 2")
+        run_identifier = lines[0]
+        uif_path = out_path(lines[1])
+
+        if not exists(uif_path):
+            raise ValueError(f"{uif_path} does not exist")
+
         sheet = load_sample_sheet(uif_path)
+
         # on Amplicon runs, lane_number is always 1, and this will be
         # properly reflected in the dummy sample-sheet as well.
         lane_number = sheet.get_lane_number()
@@ -99,12 +118,27 @@ def sequence_processing_pipeline(qclient, job_id, parameters, out_dir):
             f.write("This job was restarted.\n"
                     "failed_samples.html may contain incorrect data.\n")
     else:
+        # This is a new, fresh SPP job. Get the parameters from the user and
+        # create a 'restart_me' file in the working directory.
+
         status_line.update_current_message("Setting up pipeline")
+
         # available fields for parameters are:
         #   run_identifier, sample_sheet, content_type, filename, lane_number
+
         run_identifier = parameters.pop('run_identifier')
         user_input_file = parameters.pop('sample_sheet')
         lane_number = parameters.pop('lane_number')
+
+        # the run_identifier must be saved because it is not always preserved
+        # in a dependable location downstream. The user input file must be
+        # saved because it is always a unique name and it cannot be guaranteed
+        # to be the only .csv or .txt file at the root level of the working
+        # directory. The Lane number is not needed because it is written into
+        # the user_input file on the first run.
+        restart_file_path = out_path('restart_me')
+        with open(restart_file_path, 'w') as f:
+            f.write(f"{run_identifier}\n{user_input_file}")
 
         if {'body', 'content_type', 'filename'} != set(user_input_file):
             return False, None, ("This doesn't appear to be a valid sample "
@@ -121,7 +155,7 @@ def sequence_processing_pipeline(qclient, job_id, parameters, out_dir):
         kwargs = {'qclient': qclient,
                   'uif_path': uif_path,
                   'lane_number': lane_number,
-                  'config_path': CONFIG_FP,
+                  'config_fp': CONFIG_FP,
                   'run_identifier': run_identifier,
                   'output_dir': out_dir,
                   'job_id': job_id,
